add `Iterators` module around extra iterator functionality
